telegram-bot-v2/requirements.txt
aiogram==3.4.0
asyncpg==0.29.0
python-dotenv==1.0.0
